Progress update — I implemented the preferred Option A (rename underscored entity fields to camelCase + add @Column mappings) across the codebase where necessary, updated DTO mappers, repositories, services, and controllers that used the old names

+++ UnifiedPatientManager/src/main/java/com/Eges411Team/UnifiedPatientManager/DTOs/requests/AllergyRequest.java
@@ -10,10 +10,10 @@ public class AllergyRequest {
 
     // optional: let client send id if needed (e.g. for updates)
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
 
     @JsonProperty("patient_id")
-    private Integer patientId;
+    private Long patientId;
 
     @JsonProperty("reaction")
     private String reaction;
